+++ qson/deserialize/string.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
-#include "qson.h"
-#include "_qson.h"
+#include <qson/qson.h>
+#include "_deserialize.h"
 
 inline static qson_result handle_escape(qson_deserialize_ctx_t *ctx, char *buffer, int *sizep, int *ip) {
 	int size = *sizep;
