+++ test/test.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 #include <qson/qson.h>
-#include <qson/_qson.h>
+#include <qson/deserialize/_deserialize.h>
 
 #ifdef __cplusplus
 extern "C" {
